Move filter to predicate pkg since it's actually a predicate
